feat: 위키 상세 페이지 디자인 고도화 및 기능 보완 완료 (#214)
* feat: 위키 상세 페이지 디자인 고도화 및 기능 보완 완료

* feat: 개인정보취급방침 링크를 새탭에서 열리도록 수정

+++ src/_components/MaeilWiki/Wiki/WikiQuestion.tsx
@@ -5,6 +5,8 @@ import { Member, Wiki } from '../_types/wiki';
 import { WIKI_CATEGORY_KO } from '../_constants/wikiCategory';
 import {
   anonymousProfileImage,
+  deleteButton,
+  ownerMark,
   profileImage,
   profileLink,
   profileName,
@@ -12,11 +14,16 @@ import {
   wikiCreatedAt,
   wikiOwnerProfile,
   wikiQuestion,
+  wikiQuestionDetail,
   wikiQuestionHeader,
   wikiQuestionInfo,
+  wikiQuestionTitle,
 } from './wiki.css';
 import MDPreview from '../Markdown/MDPreview';
 import { useDeleteWiki } from './_hooks/useDeleteWiki';
+import { calculateElapsedTime } from '../_utils/calculateElapsedTime';
+import { useAuthContext } from '../_store/authContext';
+import GenericAvatar from '@/_assets/images/maeilWiki/genericAvatar.svg';
 
 export interface WikiQuestionProps {
   wiki: Wiki;
@@ -25,24 +32,45 @@ export interface WikiQuestionProps {
 export default function WikiQuestion({ wiki }: WikiQuestionProps) {
   const { mutate: deleteWiki } = useDeleteWiki(wiki.id);
 
+  const { user } = useAuthContext();
+  const isOwner = user?.id === wiki.owner.id;
+
   const onClickDeleteWiki = () => {
+    if (!isOwner) {
+      return;
+    }
+
     if (confirm('질문을 삭제하시겠습니까?')) {
       deleteWiki();
     }
   };
 
+  const { category, question, owner, createdAt, questionDetail } = wiki;
+
   return (
     <header className={wikiQuestionHeader}>
-      <div className={wikiQuestionInfo}>
-        <WikiOwnerProfile owner={wiki.owner} />
-        <p className={wikiCreatedAt}>{wiki.createdAt}</p>
+      <div className={wikiQuestionTitle}>
+        <p className={wikiCategory({ category })}>{WIKI_CATEGORY_KO[category]}</p>
+        <h1 className={wikiQuestion}>{question}</h1>
       </div>
-      <div>
-        <h1 className={wikiQuestion}>Q. {wiki.question}</h1>
-        <p className={wikiCategory}>[분야: {WIKI_CATEGORY_KO[wiki.category]}]</p>
+      <div className={wikiQuestionDetail}>
+        <div className={wikiQuestionInfo}>
+          <WikiOwnerProfile owner={owner} />
+          <span className={ownerMark}>작성자</span>
+          <div className={wikiCreatedAt}>
+            <div>{calculateElapsedTime(createdAt)}</div>
+            {isOwner && (
+              <>
+                <div>•</div>
+                <div role="button" onClick={onClickDeleteWiki} className={deleteButton}>
+                  삭제
+                </div>
+              </>
+            )}
+          </div>
+        </div>
+        {questionDetail && <MDPreview source={questionDetail} />}
       </div>
-      {wiki.questionDetail && <MDPreview source={wiki.questionDetail} />}
-      <button onClick={onClickDeleteWiki}>삭제</button>
     </header>
   );
 }
@@ -51,14 +79,14 @@ function WikiOwnerProfile({ owner }: { owner: Member }) {
   if (!owner.name) {
     return (
       <div className={wikiOwnerProfile}>
-        <div className={anonymousProfileImage} />
+        <GenericAvatar className={anonymousProfileImage} />
         <p className={profileName}>익명</p>
       </div>
     );
   }
 
   return (
-    <a href={owner.github} className={profileLink} target="_blank" rel="noopener noreferrer">
+    <a href={owner.github} className={profileLink} rel="noopener noreferrer">
       <div className={wikiOwnerProfile}>
         <img className={profileImage} src={owner.profileImage} />
         <p className={profileName}>{owner.name}</p>
